Move TestHelpers module into own file
